Game: Code cleanup
