+++ game/src/game/Entities/Object.hpp
@@ -809,6 +809,8 @@ public:
 
     void dropAllItems();
 
+    const oglx_texture_t* getIcon() const;
+
 private:
 
     /**
